scripts/requirements.txt
paddlepaddle==2.6.1
paddleocr==2.7.0
Pillow==10.2.0
